pyproject.toml
[tool.poetry]
name = "celai-chatwoot"
version = "0.1.0"
description = ""
authors = ["Alex Martin <alejamp@gmail.com>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "celai_chatwoot" }
]

[tool.poetry.urls]
Homepage = "https://github.com/cel-ai/celai"
Issues = "https://github.com/cel-ai/celai/issues"

[tool.poetry.dependencies]
#Runtime
python = "^3.11"
# Dependencies for the project:
woot =  { git = "https://github.com/dearkafka/woot.git", rev = "3be2156" }
celai = "^0.1.14"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]